# swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "NeuroSuitUp API",
        "description": "Tools for communicating with the NeuroSuitUp database",
        "version": "1.0.0",
        "contact": {
        "name": "Apostolos Praftsiotis",
        "email": "apospraf@gmail.com"
        }
    },
    "servers": [
        {
            "url":"http://localhost:3000",
            "description":"Temp localhost server"
        },
        {
            "url": "https://neurosuituptest.herokuapp.com/",
            "description": "Heroku free server"
        }, 
        {
            "url": "http://192.168.0.188:3000",
            "description": "Heroku free server"
        }
    ],
    "components": {
      "schemas": { 
...
        "TaskInsert": {
            "type": "object",
            "properties": {
                "codename": {
                    "type": "string",
                    "description": "The codename of the specific task"
                },
                "isPublic": {
                    "type": "boolean",
                    "description": "Whether or not the specific task is available to every instructor"
                },
                "difficulty": {
                    "type": "integer",
                    "description": "The difficulty of the specific task (1 - 10)"
                },
                "poses": {
                    "type": "array",
                    "description": "Array containing the poses that define the specific task",
                    "items": {
                        "type": "string",
                        "description": "The pose id (ObjectId mongo)"
                    }
                }
            },
            "required": ["codename", "isPublic", "difficulty", "poses"]
        },
        "Task": {
          "allOf": [
              {"$ref": "#/components/schemas/TaskInsert"},
              {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "The task id (ObjectId, mongodb)"
                    }
                }
              }
          ]  
        },
        "ExerciseInsert": {
            "type": "object",
            "properties": {
                "codename": {
                    "type": "string",
                    "description": "The codename of the specific exercise"
                },
                "isPublic": {
                    "type": "boolean",
                    "description": "Whether or not the specific exercise is available to every instructor"
                },
                "tasks": {
                    "type": "array",
                    "description": "Array containing the tasks the define the specifc exercise",
                    "items":{
                        "type": "string",
                        "description": "The task id (ObjectId, mongodb)"
                    }
                },
                "delays": {
                    "type": "array",
                    "description": "The delay in seconds for each task pair",
                    "items": {
                        "type": "string",
                        "description": "The delay in seconds (sec)"
                    }
                }
            }
        },
        "Exercise": {
            "allOf": [
                {"$ref": "#/components/schemas/ExerciseInsert"},
                {
                    "type": "object",
                    "properties": {
                       "_id": {
                           "type": "string",
                           "description": "The exercise id (ObjectId, mongodb)"
                        }
                    }
                }
            ]
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }, 
        "Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
      },
      "parameters": {
          "selectstring": {
              "in": "header",
              "name": "selectstring",
              "description": "Selects the fields of the returned resource",
              "schema": {
                  "type": "string"
              },
              "allowEmptyValue": true
          }
      },
      "responses": {
...
        "Unauthorized": {
            "description": "Authorization failed, either by unsuccessfull verification of token, or by limited permissions",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }, 
        "InternalError": {
            "description": "Internal or structural error",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/tasks": {
            "get": {
                "summary": "Get all available tasks for the logged in instructor",
                "tags": ["tasks"],
                "parameters": [{
                    "$ref": "#/components/parameters/selectstring"
                }], 
                "responses": {
                    "200": {
                        "description": "Query successfull, return all found tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "integer"
                                        },
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                }
            },
            "post": {
                "summary": "Insert a new task to the database",
                "tags": ["tasks"],
                "requestBody": {
                    "description": "The new task",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskInsert"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task successfully inserted to the database",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "409": {
                        "description": "Task codename already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalError"
                    }
                }
            }  
      },
      "/tasks/{taskId}": {
          "parameters": [{
              "in": "path",
              "name": "taskId",
              "description": "The requested task id",
              "required": true,
              "schema": {
                  "type": "string"
              }
          }],
          "get": {
              "summary": "Get task by id",
              "tags": ["tasks"],
              "parameters": [{
                  "$ref": "#/components/parameters/selectstring"
              }],
              "responses": {
                  "200": {
                      "description": "Query was successful, return found task",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "object",
                                  "properties": {
                                      "task": {
                                          "$ref": "#/components/schemas/Task"
                                      }
                                  }
                              }
                          }
                      }
                  }, 
                  "404": {
                      "description": "Task was not found",
                      "content": {
                          "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Message"
                              }
                          }
                      }
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          },
          "delete": {
              "summary": "Delete task by id",
              "tags": ["tasks"],
              "responses": {
                  "200": {
                      "description": "Successfully deleted task",
                      "content": {
                          "application/json": {
                              "schema": {
                                "$ref": "#/components/schemas/Message"
                              }
                          }
                      }
                  },
                  "401": {
                      "$ref": "#/components/responses/Unauthorized"
                  },
                  "403": {
                    "description": "Task is not owned, thus cannot be deleted",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                  },
                  "409": {
                      "description": "Deleting this task will affect some Exercises",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "object",
                                  "properties": {
                                      "message": {
                                          "type": "string"
                                      },
                                      "exercises": {
                                          "$ref": "#/components/schemas/Exercise"
                                      }
                                  }
                              }
                          }
                      }
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          }
      },
      "/fetchTasks": {
        "post": {
            "summary": "Fetch specific tasks with ids, located at the body of the request",
            "tags": ["tasks"],
            "parameters": [
                {
                    "$ref": "#/components/parameters/selectstring"
                }
            ],
            "requestBody": {
                "description": "The array of the requested task ids",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "description": "The task id (ObjectId mongo)"
                            }
                        },
                        "example": "{['615c94503b3ffd5cccec496a', '615c945e3b3ffd5cccec496b']}"
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Successfully fetched all tasks",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "count": {
                                        "type": "integer",
                                        "description": "The number of found tasks"
                                    },
                                    "tasks": {
                                        "type": "array",
                                        "description": "Array containg the task ids",
                                        "items": {
                                            "type": "string"
                                          }
                                    }
                                }
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
...
    }
}